+++ config/deps.ts
@@ -2,18 +2,18 @@ export {
   objectEntries,
   objectKeys,
 } from "https://cdn.skypack.dev/ts-extras@0.11.0?dts";
-export { parse } from "https://deno.land/std@0.155.0/flags/mod.ts";
-export * from "https://deno.land/std@0.155.0/fs/mod.ts";
-export * from "https://deno.land/std@0.155.0/fs/mod.ts";
-export { readLines } from "https://deno.land/std@0.155.0/io/mod.ts";
-export { getLogger } from "https://deno.land/std@0.155.0/log/mod.ts";
+export { parse } from "https://deno.land/std@0.156.0/flags/mod.ts";
+export * from "https://deno.land/std@0.156.0/fs/mod.ts";
+export * from "https://deno.land/std@0.156.0/fs/mod.ts";
+export { readLines } from "https://deno.land/std@0.156.0/io/mod.ts";
+export { getLogger } from "https://deno.land/std@0.156.0/log/mod.ts";
+export * as semver from "https://deno.land/std@0.156.0/semver/mod.ts";
 export { globber } from "https://deno.land/x/globber@0.1.0/mod.ts";
 export {
   html,
   type Token,
   tokens,
 } from "https://deno.land/x/rusty_markdown@v0.4.1/mod.ts";
-export * as semver from "https://deno.land/x/semver@v1.4.1/mod.ts";
 export { udd, type UddResult } from "https://deno.land/x/udd@0.7.5/mod.ts";
 export { default as gitUrlParse } from "https://esm.sh/git-url-parse@12.0.0";
 export { default as pLimit } from "https://esm.sh/p-limit@4.0.0";